requirements.txt
langchain-openai>=0.1.0
langgraph>=0.0.60
langchain-core>=0.2.0
tavily-python>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0